requirements.txt
boto3>=1.38.0
flask>=3.0.0
flask-cors>=6.0.0
python-dotenv>=1.0.0
pillow>=11.0.0
pdf2image>=1.17.0
pytesseract>=0.3.13
streamlit>=1.45.0
streamlit-extras>=0.6.0
streamlit-chat>=0.1.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
strands-agents>=0.1.0
